There's some bitshifting error associated with how the 32 bit ints are packed in ESP. Using memcpy() seems to fix it. TODO change all such in ESP to memcpys

+++ python/ESP/espmodule.cpp
@@ -7,7 +7,7 @@
 
 /** Pack data into a GPS message */
 static PyObject * pack_gps(PyObject *self, PyObject *args){
-  long lat=0,lon=0;
+  int32_t lat=0,lon=0;
   float t = 0.0;
   if (!PyArg_ParseTuple(args,"llf",&lon,&lat,&t)){
     return NULL;
